Add parameter to set GPS message frame_id.
This commit adds a string parameter named 'frame_id' that will be
copied into the header of every gps message.  This is intended to
correspond to the GPS antenna coordinate frame so that other nodes can
determine where the measurement was taken.  The frame_id defaults to
empty to be compatible with previous behavior.