Adds configuration parameter span_frame_to_ros_frame.
This configuration parameter will apply a VEHICLEBODYROTATION command
which rotates the Novatel SPAN frame 90 degrees counter-clockwise
about the Z axis and an APPLYVEHICLEBODYROTATION command to make
that rotation apply to the IMU and INS outputs. This converts the
Novatel SPAN frame (right-hand rule with Y forward) to a ROS frame
(right-hand rule with X forward).