Use GPGGA quality to set GPSFix status
This adds enums to the Gpgga message to represent all of the different
possible qualities, and it will also use the quality of GPGGA logs in
order to set the status of the GPSFix message.  It also adds a unit
test to verify that GPGGA logs are being parsed properly.